+++ src/Translator.php
@@ -15,14 +15,14 @@ use Yiisoft\Translator\Event\MissingTranslationEvent;
  */
 final class Translator implements TranslatorInterface
 {
+    private MessageFormatterInterface $defaultMessageFormatter;
+
     private string $defaultCategory = 'app';
 
     /**
      * @var CategorySource[][] Array of category message sources indexed by category names.
      */
     private array $categorySources = [];
-
-    private ?SimpleMessageFormatter $simpleMessageFormatter = null;
 
     /**
      * @psalm-var array<string,true>
@@ -37,8 +37,10 @@ final class Translator implements TranslatorInterface
     public function __construct(
         private string $locale = 'en_US',
         private ?string $fallbackLocale = null,
-        private ?EventDispatcherInterface $eventDispatcher = null
+        private ?EventDispatcherInterface $eventDispatcher = null,
+        ?MessageFormatterInterface $defaultMessageFormatter = null,
     ) {
+        $this->defaultMessageFormatter = $defaultMessageFormatter ?? new NullMessageFormatter();
     }
 
     public function addCategorySource(CategorySource $category): void
@@ -79,7 +81,7 @@ final class Translator implements TranslatorInterface
 
         if (empty($this->categorySources[$category])) {
             $this->dispatchMissingTranslationCategoryEvent($category);
-            return $this->getSimpleMessageFormatter()->format($id, $parameters);
+            return $this->defaultMessageFormatter->format($id, $parameters, $locale);
         }
 
         return $this->translateUsingCategorySources($id, $parameters, $category, $locale);
@@ -152,13 +154,5 @@ final class Translator implements TranslatorInterface
             $this->dispatchedMissingTranslationCategoryEvents[$category] = true;
             $this->eventDispatcher->dispatch(new MissingTranslationCategoryEvent($category));
         }
-    }
-
-    private function getSimpleMessageFormatter(): SimpleMessageFormatter
-    {
-        if ($this->simpleMessageFormatter === null) {
-            $this->simpleMessageFormatter = new SimpleMessageFormatter();
-        }
-        return $this->simpleMessageFormatter;
     }
 }